+++ +sbfsem/+io/JSON.m
@@ -4,15 +4,28 @@ classdef JSON < handle
     % Description:
     %   Handles conversion of SBFSEM-tools data to JSON files.
     %
+    % Constructor:
+    %   obj = sbfsem.io.JSON(source, fPath);
+    %
+    % Input:
+    %   source      Volume name or abbreviation
+    %   fPath       File path to save JSON files (optional, if not 
+    %               specified, opens a user interface to choose directory)
     % ---------------------------------------------------------------------
     
     properties (Access = private)
         fPath
-        hasMetadata       
+        source
+        hasMetadata
     end
     
+    properties (Constant = true, Access = private)
+        NEURON_NAME = '%s%u.json';
+        METADATA_NAME = '%s_metadata.json';
+    end
+
     methods
-        function obj = JSON(fPath)
+        function obj = JSON(source, fPath)
             % JSON Constructor
             % Input:
             %   fPath   Directory to save JSON data
@@ -22,11 +35,48 @@ classdef JSON < handle
             else
                 obj.fPath = fPath;
             end
+            cd(obj.fPath);
+            
+            obj.source = validateSource(source);
+            
+            obj.hasMetadata = obj.findMetadata;
+            if ~obj.hasMetadata
+                fprintf('No metadata found for %s\n', obj.source);
+                value = questdlg(...
+                    'No volume metadata found. Create one?',...
+                    'Volume Metadata Dialog', 'Yes', 'No', 'Yes');
+                if ~isempty(value)
+                    switch value
+                        case 'Yes'
+                            obj.saveMetadata();
+                        case 'No'
+                            return;
+                    end
+                end
+            end
         end
-        
+            
         function export(obj, neuron)
-            % EXPORT
+            % EXPORT  Exports a neuron to filePath
+            % 
+            % Input:
+            %   neuron      Neuron object
+            % -------------------------------------------------------------
+            
             assert(isa(neuron, 'Neuron'), 'Input a Neuron!'); 
+            
+            fName = sprintf('%s%u.json',... 
+                getVolumeAbbrev(obj.source), num2str(neuron.ID), '.json');
+            
+            hasNeuron = obj.findNeuron(obj, ID);
+            
+            if hasNeuron
+                value = questdlg('Overwrite existing file?',...
+                    'Overwrite dialog', 'Yes', 'No', 'Yes');
+                if ~isempty(value) || strcmp(value, 'No')
+                    return;
+                end
+            end
             
             % Ensure synapses and geometries are present, if existing
             if isempty(neuron.synapses)
@@ -42,6 +92,38 @@ classdef JSON < handle
             S = struct(neuron);
             
             S = obj.prep(S);
+            
+            obj.findNeuron(fName);
+            
+            savejson('', S, [obj.fPath, filesep, fName]);
+        end
+    end
+    
+    methods (Access = private)        
+        function tf = findNeuron(obj, ID)
+            % FINDNEURON
+            str = sprintf(obj.NEURON_NAME, getVolumeAbbrev(obj.source), ID);
+            cd(obj.fPath);
+            x = cellstr(ls(str));
+            tf = ~isempty(x);
+        end
+      
+        function tf = findMetadata(obj)
+            % FINDMETADATA
+            
+            str = sprintf(obj.METADATA_NAME, obj.source);
+            cd(obj.fPath)
+            x = cellstr(ls(str));
+            tf = ~isempty(x);           
+        end
+        
+        function saveMetadata(obj)
+            % SAVEMETADATA
+            
+            S = struct('VolumeName', obj.source, 'DateCreated', datestr(now));
+            
+            dataDir = [fileparts(mfilename('fullpath')), '\data'];
+            
         end
     end
     
@@ -52,7 +134,7 @@ classdef JSON < handle
             % Remove unecessary, transient/dependent properties
             S = rmfield(S, {'ODataClient', 'GeometryClient', 'SynapseClient'});
             S = rmfield(S, {'somaRow', 'offEdges'});
-            S = rmfield(S, 'includeSynapses');
+            S = rmfield(S, {'includeSynapses', 'USETRANSFORM'});
             
             S.volumeScale = struct(...
                 'value', S.volumeScale,...
@@ -62,8 +144,10 @@ classdef JSON < handle
             S.edges = table2struct(S.edges);
             
             if ~isempty(S.synapses)
-                S.synapses.LocalName = arrayfun(@char, S.synapses.LocalName);
+                S.synapses.LocalName = arrayfun(@char, S.synapses.LocalName,...
+                    'UniformOutput', false);
             end
+            S.synapses = table2struct(S.synapses);
             
             if ~isempty(S.analysis)
                 % Create a struct for storing analyses